Added config setting for removing errors in frontend
This pr implements a simple way to remove all errors from the frontend.
The change is requested because when working with a lot of tubes that changes from being shown to not being shown, sometimes would throw a lot of errors.

This is essentially the first of three pr's related to fixing this issue.

This is my first contribution pr, so please do say if i am missing anything or have forgotten to add something :)