Adding reuse of already fetched tubes
When getting tubeStats the tubes has already been fetched right before.
As the tubeStats (in setup) sometimes failes to connect and thereby fails in receiving the excpected data, this request is proposed as an optimization to reuse the already fetched tube names if they already exists.

This is the second of three pr's to fix an issue where a lot of errors where being thrown when having a lot of tubes are not always being run by a process.

This is my first set of contribution pr's, so please do say if i am missing anything or have forgotten to add something :)