Adding retry settings for beanstalk data fetch
This pr request to implement some new options for making multiple retries upon server exceptions.

These changes are request as the server setup i work on in my day job runs a lot of tubes which sometimes is not active, which makes it so that the tube is not existing anymore. Sometimes connections will fail for just a very breaf moment, so the option to add some simple retries will not be notisable when using the console but will help a lot.

This is the third pr to fix a lot of issues i see on a daily basis in our setup when working with this console.

3 things will be requested added with this pr.

1. Adding new config options and code to making data retry possible and tweak the settings of how often and how many times
2. Storing statsTube data in the session mostly for reuse when auto refreshing and statsTube sometimes can be missing because of server exceptions
3. Some minor bug fixes and optimizations:
    3.1 checking $this->_globalVar['action'] instead of $_GET['action'] as these two should be the same, and if i remember corret $_GET removes data upon first retrieaval so it might not exist here
    3.2 $tubeStats['pause-time-left'] would sometimes be missing when auto refreshing

This is my first set of contribution pr's, so please do say if i am missing anything or have forgotten to add something :)